webapi/tct-application-tizen-tests/application/ApplicationControl_constructor.html

<!DOCTYPE html>

<!--

Copyright (c) 2013 Samsung Electronics Co., Ltd.

Licensed under the Apache License, Version 2.0 (the License);
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.



Authors:
        Mariusz Polasinski <m.polasinski@samsung.com>


-->

<html>

<head>
<title>ApplicationControl_constructor</title>
<meta charset="utf-8">
<script type="text/javascript" src="../resources/testharness.js"></script>
<script type="text/javascript" src="../resources/testharnessreport.js"></script>
</head>

<body>
<div id="log"></div>
<script type="text/javascript">

//==== TEST: ApplicationControl_constructor
//==== LABEL Test whether constructor with all arguments are supported
//==== SPEC Tizen Web API:Application:Application:ApplicationControl:constructor C
//==== SPEC_URL https://developer.tizen.org/help/topic/org.tizen.web.device.apireference/tizen/application.html
//==== PRIORITY P1
//==== TEST_CRITERIA CONSTRM CONSTRA

test(function () {
    var appControl, operation = "http://tizen.org/appcontrol/operation/share",
        uri = "share.html", mime = "image/*", category = "dummyValue", i,
        data = [new tizen.ApplicationControlData("images", ["imagedata1", "imagedata2"])];

    appControl = new tizen.ApplicationControl(operation, uri, mime, category, data);
    assert_true(appControl instanceof tizen.ApplicationControl, "object was not created properly");
    assert_equals(appControl.operation, operation,
        "the operation argument passed as constructor parameter is different than the operation attribute in the object");
    assert_equals(appControl.uri, uri, "the uri argument passed as constructor parameter is different than the uri attribute in the object");
    assert_equals(appControl.mime, mime, "the mime argument passed as constructor parameter is different than the mime attribute in the object");
    assert_equals(appControl.category, category,
        "the category argument passed as constructor parameter is different than the category attribute in the object");

    for (i = 0; i < data[0].value.length; i++) {
        assert_equals(appControl.data[0].value[i], data[0].value[i], "data passed as constructor parameter is different than data in the object");
    }
}, document.title);

</script>
</body>
</html>
